+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Deprecated `Config.retryInterval`.
+
 ## 5.2.0 - 2025-08-25
 
 ### Changed