+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 5.4.0 - 2025-11-01
+
 ### Changed
 
 - AI Guard: updated various models around the multimodal guard.